Added routed PCB
I'm adding this to the repo because I spent a lot of time on it and it
helped me get a good idea for how much this board will cost. However, it
is NOT SUITABLE FOR PRODUCTION because:

 - Every trace is 5 mils wide, even power traces, and I'm not sure that
   they will be wide enough for the current drawn.
 - The STM32, its crystal, and its programming header take up an
   obscenely large portion of the board.
 - All the current measurement/voltage injection headers take up
   precious space as well.
 - I thought the power and debug LEDs would be useful, but they take up
   too much space to be practical. I'd like to replace the power LEDs
   with test points, and either get rid of the debug LEDs entirely or
   leave one or two in there for basic status/ready/error information.
 - I routed the HDMI->Bridge traces before almost everything else, not
   realizing that it's actually the Bridge->LCD connections that put the
   greatest constraint on where everything else needs to be placed.
 - The footprints I designed are missing important things like the
   courtyard dimensions, silkscreen, etc.
 - When connected to the LCD, the board should approximately fit within
   the dimensions of the screen. The current version is not sized for
   that.
 - Some components need to be in different packages to be more
   cost-effective.

Current cost estimates from MacroFab are between ~$150 and ~$180 for a
completely assembled board, which I feel is too much to for one that
probably won't work (for the reasons stated above). After making the
necessary changes, including reducing board area and the number of
drilled holes, the cost should be much more manageable.